Refactor llm summariser and add a test.
